refactor(redbox): moving OpenStackFrameTask out of content view to enable red box migration
